negative timeline...

+++ philo/headers/utils.h
@@ -5,7 +5,7 @@
 
 int	ft_isdigit(int c);
 int	is_in_range_long(long value, long min, long max);
-long long get_time(t_data *data);
-void	ft_usleep(t_data *data, long long starting_point, int sleeping_time);
+unsigned long long get_time(t_data *data);
+void	ft_usleep(t_data *data, uint64_t starting_point, uint64_t sleeping_time);
 
 #endif
